+++ src/data/data_template.py
@@ -2,7 +2,7 @@ from src.data.data_abstract import DataAbstract
 from src.data.data_non_linear import DataNonLinear
 
 """
-Template for data classes in the KSU project. 
+Template for data classes in the KSU project.
   * Copy into new file
   * Rename class
   * Implement TODO's below
